chore: rename logger package
