+++ logger/logger.go
@@ -1,4 +1,4 @@
-package http
+package logger
 
 import (
 	"log"
